+++ WorldGenerator/World/Generator/Noise/ScaleOffset.cs
@@ -11,22 +11,22 @@ namespace Sean.WorldGenerator.Noise
         protected DataSource m_scale { get; set; }
         protected DataSource m_offset { get; set; }
 
-        public override float get (float x, float y)
+        public override double get (double x, double y)
         {
             return m_source.get (x, y) * m_scale.get (x, y) + m_offset.get (x, y);
         }
 
-        public override float get (float x, float y, float z)
+        public override double get (double x, double y, double z)
         {
             return m_source.get (x, y, z) * m_scale.get (x, y, z) + m_offset.get (x, y, z);
         }
 
-        public override float get (float x, float y, float z, float w)
+        public override double get (double x, double y, double z, double w)
         {
             return m_source.get (x, y, z, w) * m_scale.get (x, y, z, w) + m_offset.get (x, y, z, w);
         }
 
-        public override float get (float x, float y, float z, float w, float u, float v)
+        public override double get (double x, double y, double z, double w, double u, double v)
         {
             return m_source.get (x, y, z, w, u, v) * m_scale.get (x, y, z, w, u, v) + m_offset.get(x,y,z,w,u,v);
         }
